requirements.txt

PyQt6==6.6.1
cryptography==42.0.0
argon2-cffi==23.1.0
pyperclip==1.8.2
requests==2.31.0
python-dotenv==1.0.1
pyinstaller==6.4.0
